face_id/requirements.txt
opencv-python
python-facepp
PySimpleGUI
pysftp
tk
imutils
psycopg2-binary
Pillow